requirements.txt
pandas
streamlit
sqlalchemy
plotly
openpyxl
python-dotenv
